+++ libs/regexp.ts
@@ -4,7 +4,7 @@ import { extendMethod } from "./extDefine";
 
 extendMethod(RegExp, "escape", function(str: string): string {
 
-    return str.replace(/([\~\!\$\^\*\(\)\{\}\[\]\\\|\:\/\?\<\>\.])/g, "\\$1");
+    return str.replace(/([\*\.\?\+\$\^\[\]\(\)\{\}\|\\])/g, "\\$1");
 
 });