+++ sample.ts
@@ -44,9 +44,13 @@ if (new RegExp(RegExp.escape("~!")).exec("~!")) {
     console.log("Regular Expression escaped successfully.");
 }
 
-let x: number[] = [1, 2, 3, 4, 5, 4, 3, 2, 1];
+let x: number[] = [1, 2, 3, 4, 25, 4, 3, 2, 1, 43, 54, 22, 33];
 
 x.remove(1);
+
+console.log(x);
+
+x.removeByIndexes([1, 3, 6]);
 
 console.log(x);